feat(utils): add get nfts by account with sequencer
