+++ src/lib/pages/nft-collections/components/CollectionList.tsx
@@ -3,7 +3,7 @@ import { GridItem, SimpleGrid } from "@chakra-ui/react";
 import { AppLink } from "lib/components/AppLink";
 import { Loading } from "lib/components/Loading";
 import { EmptyState, ErrorFetching } from "lib/components/state";
-import type { Collection } from "lib/services/nft";
+import type { Collection } from "lib/services/types";
 import type { Option } from "lib/types";
 
 import { CollectionCard } from "./CollectionCard";
