+++ src/lib/services/nft/gql.ts
@@ -8,6 +8,16 @@ import {
   getNftsExpressionOld,
 } from "../expression";
 import {
+  zMetadata,
+  zNft,
+  zNftByNftAddressResponse,
+  zNftMintInfoResponse,
+  zNftMutateEventsResponseItem,
+  zNftOld,
+  zNftsByAccountResponse,
+  zNftTransactionsResponse,
+} from "../types/nft";
+import {
   getNftMintInfoQuery,
   getNftMutateEventsCountQuery,
   getNftMutateEventsCountQueryOld,
@@ -26,58 +36,8 @@ import {
   getNftTransactionsQuery,
   getNftTransactionsQueryOld,
 } from "lib/query";
-import { zBechAddr, zHexAddr, zHexAddr32, zRemark, zUtcDate } from "lib/types";
-import type { HexAddr, HexAddr32, MutateEvent } from "lib/types";
-import { parseTxHash, parseWithError, snakeToCamel } from "lib/utils";
-
-const zNft = z
-  .object({
-    uri: z.string(),
-    token_id: z.string(),
-    description: z.string().optional(),
-    is_burned: z.boolean(),
-    owner: zHexAddr,
-    id: zHexAddr32,
-    collection: zHexAddr32,
-    collectionByCollection: z.object({
-      name: z.string(),
-    }),
-  })
-  .transform((val) => ({
-    uri: val.uri,
-    tokenId: val.token_id,
-    description: val.description,
-    isBurned: val.is_burned,
-    ownerAddress: val.owner,
-    nftAddress: val.id,
-    collectionAddress: val.collection,
-    collectionName: val.collectionByCollection.name,
-  }));
-export type Nft = z.infer<typeof zNft>;
-
-const zNftOld = z
-  .object({
-    uri: z.string(),
-    token_id: z.string(),
-    description: z.string().optional(),
-    is_burned: z.boolean(),
-    vmAddressByOwner: z.object({ vm_address: zHexAddr }),
-    vm_address: z.object({ vm_address: zHexAddr32 }).optional(),
-    collectionByCollection: z.object({
-      vm_address: z.object({ vm_address: zHexAddr32 }),
-      name: z.string(),
-    }),
-  })
-  .transform<Nft>((val) => ({
-    uri: val.uri,
-    tokenId: val.token_id,
-    description: val.description,
-    isBurned: val.is_burned,
-    ownerAddress: val.vmAddressByOwner?.vm_address,
-    nftAddress: val.vm_address?.vm_address || ("" as HexAddr32),
-    collectionAddress: val.collectionByCollection.vm_address.vm_address,
-    collectionName: val.collectionByCollection.name,
-  }));
+import type { HexAddr, HexAddr32 } from "lib/types";
+import { parseWithError } from "lib/utils";
 
 export const getNfts = async (
   indexer: string,
@@ -108,11 +68,6 @@ export const getNfts = async (
     return parseWithError(zNftOld.array(), res.data.data.nfts);
   }
 };
-
-const zNftByNftAddressResponse = z.object({
-  data: z.union([zNft, zNftOld]).optional(),
-});
-export type NftByNftAddressResponse = z.infer<typeof zNftByNftAddressResponse>;
 
 export const getNftByNftAddress = async (
   indexer: string,
@@ -157,26 +112,6 @@ export const getNftsCountByAccount = async (
   }
 };
 
-const zNftMintInfoResponse = z
-  .object({
-    account: z.object({
-      address: zBechAddr,
-    }),
-    block: z.object({
-      timestamp: zUtcDate,
-      height: z.number(),
-    }),
-    hash: z.string().transform(parseTxHash),
-  })
-  .transform((val) => ({
-    minter: val.account.address,
-    txhash: val.hash,
-    height: val.block.height,
-    timestamp: val.block.timestamp,
-  }));
-export type NftMintInfo = z.infer<typeof zNftMintInfoResponse>;
-
-// TODO: combine with getNftByNftAddress when migrating to API
 export const getNftMintInfo = async (indexer: string, nftAddress: HexAddr32) =>
   axios
     .post(indexer, {
@@ -190,45 +125,8 @@ export const getNftMintInfo = async (indexer: string, nftAddress: HexAddr32) =>
       )
     );
 
-export const zMetadata = z
-  .object({
-    name: z.string(),
-    description: z.string(),
-    image: z.string(),
-    image_url: z.string().optional(),
-    attributes: z
-      .object({
-        trait_type: z.string(),
-        display_type: z.string().optional(),
-        value: z.union([z.string(), z.number()]),
-      })
-      .array()
-      .optional(),
-  })
-  .transform(snakeToCamel);
-export type Metadata = z.infer<typeof zMetadata>;
-
 export const getMetadata = async (uri: string) =>
   axios.get(uri).then(({ data }) => parseWithError(zMetadata, data));
-
-const zNftTransactionsResponse = z
-  .object({
-    transaction: z.object({
-      hash: z.string().transform(parseTxHash),
-      block: z.object({ timestamp: zUtcDate, height: z.number() }),
-    }),
-    is_nft_burn: z.boolean(),
-    is_nft_mint: z.boolean(),
-    is_nft_transfer: z.boolean(),
-  })
-  .transform((val) => ({
-    txhash: val.transaction.hash,
-    timestamp: val.transaction.block.timestamp,
-    isNftBurn: val.is_nft_burn,
-    isNftMint: val.is_nft_mint,
-    isNftTransfer: val.is_nft_transfer,
-  }));
-export type NftTransactions = z.infer<typeof zNftTransactionsResponse>;
 
 export const getNftTransactions = async (
   indexer: string,
@@ -280,22 +178,6 @@ export const getNftTransactionsCount = async (
   }
 };
 
-const zNftMutateEventsResponseItem = z
-  .object({
-    old_value: z.string(),
-    new_value: z.string(),
-    mutated_field_name: z.string(),
-    remark: zRemark,
-    block: z.object({ timestamp: zUtcDate }),
-  })
-  .transform<MutateEvent>((val) => ({
-    oldValue: val.old_value,
-    newValue: val.new_value,
-    mutatedFieldName: val.mutated_field_name,
-    remark: val.remark,
-    timestamp: val.block.timestamp,
-  }));
-
 export const getNftMutateEvents = async (
   indexer: string,
   nftAddress: HexAddr32,
@@ -345,21 +227,6 @@ export const getNftMutateEventsCount = async (
       .parse(res.data.data.nft_mutation_events_aggregate.aggregate.count);
   }
 };
-
-const zNftsByAccountResponse = z
-  .object({
-    nfts: z.union([zNft, zNftOld]).array(),
-    nfts_aggregate: z.object({
-      aggregate: z.object({
-        count: z.number(),
-      }),
-    }),
-  })
-  .transform((val) => ({
-    nfts: val.nfts,
-    total: val.nfts_aggregate.aggregate.count,
-  }));
-export type NftsByAccountResponse = z.infer<typeof zNftsByAccountResponse>;
 
 export const getNftsByAccount = async (
   indexer: string,
