+++ src/lib/services/types/index.ts
@@ -10,3 +10,5 @@ export * from "./wasm";
 export * from "./tx";
 export * from "./move";
 export * from "./stats";
+export * from "./nft-collection";
+export * from "./nft";
